+++ server/cshr/celery/send_email.py
@@ -115,7 +115,9 @@ def send_email_for_vacation_request(user_id, data):
     from django.core.mail import send_mail
     from server.cshr.models.users import User
     from server.cshr.utils.send_email import get_email_recievers
-    from server.cshr.utils.emails_templates import get_vacation_request_email_template
+    from server.cshr.utils.email_messages_templates import (
+        get_vacation_request_email_template,
+    )
 
     user: User = User.objects.get(pk=user_id)
     recievers = get_email_recievers(user)
@@ -129,7 +131,9 @@ def send_email_for_hr_letter_request(user_id, data):
     from django.core.mail import send_mail
     from server.cshr.models.users import User
     from server.cshr.utils.send_email import get_email_recievers
-    from server.cshr.utils.emails_templates import get_hr_letter_request_email_template
+    from server.cshr.utils.email_messages_templates import (
+        get_hr_letter_request_email_template,
+    )
 
     user: User = User.objects.get(pk=user_id)
     recievers = get_email_recievers(user)
@@ -143,7 +147,7 @@ def send_email_for_compensation_request(user_id, data):
     from django.core.mail import send_mail
     from server.cshr.models.users import User
     from server.cshr.utils.send_email import get_email_recievers
-    from server.cshr.utils.emails_templates import (
+    from server.cshr.utils.email_messages_templates import (
         get_compensation_request_email_template,
     )
 
@@ -159,7 +163,9 @@ def send_email_for_vacation_reply(approving_user_id, data):
     from django.core.mail import send_mail
     from server.cshr.models.users import User
     from server.cshr.utils.send_email import get_email_recievers
-    from server.cshr.utils.emails_templates import get_vacation_reply_email_template
+    from server.cshr.utils.email_messages_templates import (
+        get_vacation_reply_email_template,
+    )
 
     approving_user: User = User.objects.get(pk=approving_user_id)
     applying_user_id = data["applying_user"]
@@ -175,7 +181,9 @@ def send_email_for_hr_letter_reply(approving_user_id, data):
     from django.core.mail import send_mail
     from server.cshr.models.users import User
     from server.cshr.utils.send_email import get_email_recievers
-    from server.cshr.utils.emails_templates import get_hr_letter_reply_email_template
+    from server.cshr.utils.email_messages_templates import (
+        get_hr_letter_reply_email_template,
+    )
 
     approving_user: User = User.objects.get(pk=approving_user_id)
     applying_user_id = data["applying_user"]
@@ -191,7 +199,9 @@ def send_email_for_compensation_reply(approving_user_id, data):
     from django.core.mail import send_mail
     from server.cshr.models.users import User
     from server.cshr.utils.send_email import get_email_recievers
-    from server.cshr.utils.emails_templates import get_compensation_reply_email_template
+    from server.cshr.utils.email_messages_templates import (
+        get_compensation_reply_email_template,
+    )
 
     approving_user: User = User.objects.get(pk=approving_user_id)
     applying_user_id = data["applying_user"]
